Implemented performance testing
Ive added an in browser performance test for the run method, what it
does is described in the readme. The is also a link to a google sheets
document where the outcome is plotted in a graph.

Then ive also implemented a performance test for running 1000 commands,
which draws a timeline of time passed since start after each command.